chore: Fix Circle CI failing on a missing clang lib
The ubsan build fails due to the missing
/usr/lib/llvm-18/lib/clang/18/lib/linux/libclang_rt.ubsan_standalone-x86_64.a
which is provided by the libclang-rt-18-dev package, which would be
installed if we didn't disallow recommends.

Installing it manually is not practical, as using libclang-rt-18-dev
once clang updates to version 19 would be wrong, and installing
libclang-rt-*-dev would attempt to install all of them, either failing
or installing all of the clang version available along with them. Thus
allowing apt to automatically install recommended packages seems like
best solution.